requirements.txt
appdirs==1.4.3
awsebcli==3.12.1
blessed==1.14.2
boto3==1.4.7
botocore==1.8.21
CacheControl==0.12.3
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.7
coverage==4.4.1
dj-database-url==0.4.2
Django==1.11
django-admin-bootstrapped==2.5.7
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-bootstrap3==9.0.0
django-debug-toolbar==1.8
django-extensions==1.9.1
django-sass-processor==0.5.5
django-tables2==1.12.0
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.6
iso3166==0.8
jmespath==0.9.3
libsass==0.13.2
lxml==3.8.0
msgpack-python==0.4.8
numpy==1.13.3
packaging==16.8
pathspec==0.5.0
polyline==1.3.2
psycopg2==2.7.1
pyasn1==0.4.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
semantic-version==2.5.0
six==1.10.0
sklearn==0.0
sqlparse==0.2.4
tabulate==0.7.5
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
Werkzeug==0.12.2
